+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - AS3 P-code ValueKind namespaces handling
 - AS3 direct editation - namespace definition without explicit value
 - AS3 direct editation - var/const outside package
+- AS3 interfaces - internal modifier on methods
 
 ### Changed
 - AS1/2/3 P-code - format Number values with EcmaScript toString function